refactor: extract shared MediaDetailHero component
The movie and TV detail page branches rendered near-identical hero
layouts (backdrop, rating badge, title, metadata, description, trailer).
Extract a shared MediaDetailHero component to eliminate this duplication
so visual changes only need to happen once.
